libavcodec/assenc: add "-ass_raw_mode" option to pass thru old-style ASS
accepts old-style ASS input of free format that does not conform to the
ffmpeg-specific format, to be muxed/output as-is to a single data file,
like "-f data output.ass".
Previously, subtitles input were assumed to be in the same format as
assdec decoder, and converted to the "new" format that does not contain
"Dialogue:" header and start time field.